Allows to define additional apt repositories
Predefined apt repositories are specific to our applications. We already
have an generic configuration file for our applications so those
information should be in this file not in the build chain.
