+++ grocker/__init__.py
@@ -3,7 +3,7 @@
 
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-__version__ = '3.0.1'
+__version__ = '3.1.dev2'
 __copyright__ = '2015, Polyconseil'
 __description__ = 'Docker image builder'
